# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/decks/deck": {
      "post": {
        "tags": [
          "Decks"
        ],
        "summary": "Create Deck",
        "operationId": "create_deck_api_decks_deck_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decks/deck/{deck_id}": {
      "get": {
        "tags": [
          "Decks"
        ],
        "summary": "Read Deck",
        "operationId": "read_deck_api_decks_deck__deck_id__get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Decks"
        ],
        "summary": "Update Deck",
        "operationId": "update_deck_api_decks_deck__deck_id__put",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Decks"
        ],
        "summary": "Delete Deck",
        "operationId": "delete_deck_api_decks_deck__deck_id__delete",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decks/decks": {
      "get": {
        "tags": [
          "Decks"
        ],
        "summary": "Read Decks",
        "operationId": "read_decks_api_decks_decks_get",
        "parameters": [
          {
            "name": "subject_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeckRead"
                  },
                  "title": "Response Read Decks Api Decks Decks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/deck_configs/deck_config": {
      "post": {
        "tags": [
          "Deck Configuration"
        ],
        "summary": "Create Deck Config",
        "operationId": "create_deck_config_api_deck_configs_deck_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/deck_configs/deck_config/{deck_config_id}": {
      "get": {
        "tags": [
          "Deck Configuration"
        ],
        "summary": "Read Deck Config",
        "operationId": "read_deck_config_api_deck_configs_deck_config__deck_config_id__get",
        "parameters": [
          {
            "name": "deck_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Deck Configuration"
        ],
        "summary": "Update Deck Config",
        "operationId": "update_deck_config_api_deck_configs_deck_config__deck_config_id__patch",
        "parameters": [
          {
            "name": "deck_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Deck Configuration"
        ],
        "summary": "Delete Deck Config",
        "operationId": "delete_deck_config_api_deck_configs_deck_config__deck_config_id__delete",
        "parameters": [
          {
            "name": "deck_config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Config Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/page/decks/deck/{id}": {
      "get": {
        "tags": [
          "Decks"
        ],
        "summary": "Read Deck",
        "operationId": "read_deck_page_decks_deck__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/page/deck_configs/deck_config/create": {
      "get": {
        "tags": [
          "Deck Configurations"
        ],
        "summary": "Create Deck Config Form",
        "operationId": "create_deck_config_form_page_deck_configs_deck_config_create_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Deck Configurations"
        ],
        "summary": "Create Deck Config",
        "operationId": "create_deck_config_page_deck_configs_deck_config_create_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/DeckConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/page/deck_configs/deck_config/{id}": {
      "get": {
        "tags": [
          "Deck Configurations"
        ],
        "summary": "Read Deck Config",
        "operationId": "read_deck_config_page_deck_configs_deck_config__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeckConfigCreate": {
        "properties": {
          "deck_id": {
            "type": "string",
            "format": "uuid",
            "title": "Deck Id"
          },
          "prompt_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Fields"
          },
          "answer_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answer Fields"
          },
          "prompt_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Pool"
          },
          "prompt_pool_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Prompt Pool Counts"
          },
          "answer_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answer Pool"
          },
          "answer_pool_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Answer Pool Counts"
          }
        },
        "type": "object",
        "required": [
          "deck_id",
          "prompt_fields",
          "answer_fields",
          "prompt_pool",
          "prompt_pool_counts",
          "answer_pool",
          "answer_pool_counts"
        ],
        "title": "DeckConfigCreate"
      },
      "DeckConfigRead": {
        "properties": {
          "deck_id": {
            "type": "string",
            "format": "uuid",
            "title": "Deck Id"
          },
          "prompt_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Fields"
          },
          "answer_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answer Fields"
          },
          "prompt_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prompt Pool"
          },
          "prompt_pool_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Prompt Pool Counts"
          },
          "answer_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answer Pool"
          },
          "answer_pool_counts": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Answer Pool Counts"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "deck_id",
          "prompt_fields",
          "answer_fields",
          "prompt_pool",
          "prompt_pool_counts",
          "answer_pool",
          "answer_pool_counts",
          "id"
        ],
        "title": "DeckConfigRead"
      },
      "DeckConfigUpdate": {
        "properties": {
          "deck_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deck Id"
          },
          "prompt_fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Fields"
          },
          "answer_fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Fields"
          },
          "prompt_pool": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Pool"
          },
          "prompt_pool_counts": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Pool Counts"
          },
          "answer_pool": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Pool"
          },
          "answer_pool_counts": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Pool Counts"
          }
        },
        "type": "object",
        "title": "DeckConfigUpdate"
      },
      "DeckCreate": {
        "properties": {
          "subject_id": {
            "type": "string",
            "format": "uuid",
            "title": "Subject Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "deck_schema": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Deck Schema"
          }
        },
        "type": "object",
        "required": [
          "subject_id",
          "name",
          "deck_schema"
        ],
        "title": "DeckCreate"
      },
      "DeckRead": {
        "properties": {
          "subject_id": {
            "type": "string",
            "format": "uuid",
            "title": "Subject Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "deck_schema": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Deck Schema"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "card_count": {
            "type": "integer",
            "title": "Card Count"
          }
        },
        "type": "object",
        "required": [
          "subject_id",
          "name",
          "deck_schema",
          "id",
          "card_count"
        ],
        "title": "DeckRead"
      },
      "DeckUpdate": {
        "properties": {
          "subject_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "DeckUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}